+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.84 (unpublished)
 
+- Coco.translator now does not eats all errors in loadMessages success callback, so you can use promises OR callback
+
 ## 0.1.83
 
 - fixed console.warning bug on Coco.Translator